Week4 rewrite
I'm not quite as happy with this as I was with week2, I think it could use more work. But it has been re-written, and I built the project as I wrote it. I will review it again when I build the project only following the guide I wrote, and see if it covers everything it should. I'm doubtful that it will do everything it needs, and may need to be expanded.